restructure of requests into packaged services
